requirements.txt
azure-cognitiveservices-vision-computervision==0.7.0
gradio
python-dotenv
translate
